+++ UnitTests/UnitTest_SpectrumUtils.cpp
@@ -48,7 +48,7 @@ namespace novac
             bool ret = novac::Derivative(inputSpectrum.m_data, inputSpectrum.m_length, result);
 
             REQUIRE(ret == true);
-            REQUIRE(inputSpectrum.m_length == result.size());
+            REQUIRE(inputSpectrum.m_length == static_cast<long>(result.size()));
             for (size_t ii = 0; ii < result.size(); ++ii)
             {
                 REQUIRE(std::abs(result[ii] < 1e-9));
